styled basics of pages and sidebar
- added common.css stylesheet, which is common to all pages. Includes sidebar as well as body and main styling.
You may create and use an additional pagename.css for page-specific styling (the code for it already there, just in a comment).
Do not remove common when doing so. Do not change the order of the css pages. Common needs to be called first, and pagename.css second so it can overwrite styling if needed.

- added source for background image, do not remove 
- added embed video in Home page + placeholder title